% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmfslow.R
\name{generateStart}
\alias{generateStart}
\title{Generate a (random) start}
\usage{
generateStart(n, m)
}
\arguments{
\item{n}{total number of mediators}

\item{m}{number of selected mediators}
}
\value{
binary vector of mediator selections
}
\description{
Generate a (random) start
}
\keyword{internal}
