% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{registerCatFun}
\alias{registerCatFun}
\title{Register a printing function for verbose output}
\usage{
registerCatFun(w, m)
}
\arguments{
\item{w}{half the width of the console}

\item{m}{the number of mediators}
}
\value{
callable printing function
}
\description{
Register a printing function for verbose output
}
\keyword{internal}
