% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{registerDecFun}
\alias{registerDecFun}
\title{Register a decision function}
\usage{
registerDecFun(decisionFunction)
}
\arguments{
\item{decisionFunction}{string or function}
}
\value{
decision function to be used
}
\description{
Register a decision function
}
\keyword{internal}
