% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{registerLapplyFun}
\alias{registerLapplyFun}
\title{Register an lapply function for parallel or serial processing with or
without progress bar.}
\usage{
registerLapplyFun(parallel, progressBar, nCores)
}
\arguments{
\item{parallel}{true/false}

\item{progressBar}{true/false}
}
\value{
drop-in replacement lapply function
}
\description{
Register an lapply function for parallel or serial processing with or
without progress bar.
}
\keyword{internal}
