% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update.cmf}
\alias{update.cmf}
\title{Add samples to existing cmf results object}
\usage{
\method{update}{cmf}(object, nStarts = 100, ...)
}
\arguments{
\item{object}{a cmf object}

\item{nStarts}{the number of starts to add (default 100)}

\item{...}{not used}
}
\description{
This function adds additional samples to the existing results object
}
\examples{
# generate some data
dat <- generateMed(a = (1:10)/20, b = (1:10)/20)
res <- cmf(dat)
# double the samples
res <- update(res, 1000)

}
