% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cmf-methods}
\alias{cmf-methods}
\alias{plot.cmf}
\alias{summary.cmf}
\alias{print.cmf}
\alias{setCutoff}
\alias{screeplot.cmf}
\title{Methods for cmf results lists}
\usage{
\method{plot}{cmf}(x, removeZeros = FALSE, line = TRUE, las = 2,
  ylim = c(0, 1), ...)

\method{summary}{cmf}(object, ...)

\method{print}{cmf}(x, ...)

setCutoff(object, cutoff)

\method{screeplot}{cmf}(x, topn, border = NA, space = 0, ...)
}
\arguments{
\item{x}{cmf object}

\item{removeZeros}{whether to remove the unselected mediators from the plot}

\item{line}{whether to show a line at the chosen cutoff}

\item{las}{las argument to barplot}

\item{ylim}{y limits argument to barplot}

\item{...}{other arguments passed to barplot and summary}

\item{object}{cmf object}

\item{cutoff}{new cutoff for mediator selections. If left out, use
changepoint detection to automatically select the number of variables.}

\item{topn}{only show the top n mediators}

\item{border}{the colour of the border around the bars (default NA)}

\item{space}{the amount of space between the bars (default 0)}
}
\description{
Methods for cmf results lists
}
\seealso{
\code{\link{cmf}}
}
