% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmf.R
\name{cmf}
\alias{cmf}
\title{Coordinate-wise Mediation Filter}
\usage{
cmf(x, M, y, decisionFunction = "corMinusPartCor", maxIter = 10,
  stableLag = 1, nStarts = 1, randomStart = TRUE, randomOrder = FALSE,
  subSampling = FALSE, cutoff = 0.5, parallel = FALSE, nCores = 2,
  verbose = FALSE, progressBar = FALSE, ...)
}
\arguments{
\item{x}{exogenous variable; numeric vector}

\item{M}{potential mediators; data frame with column names}

\item{y}{outcome variable; numeric vector}

\item{decisionFunction}{either a function with as inputs x, m, y, parameters,
and as output a TRUE (include) or FALSE (exclude) statement or a string
indicating the built-in decision function to use (see details)}

\item{maxIter}{the maximum number of iterations for each start}

\item{stableLag}{how long does the selection need to be stable before
deciding upon convergence}

\item{nStarts}{how many times to start the algorithm, combine this with the
randomStart parameter}

\item{randomStart}{either FALSE for a start with no variables selected, or
the number of mediators to be randomly drawn for the initiations. TRUE means
half of the available degrees of freedom will be spent or half the mediators
will be selected; whichever is lowest.}

\item{randomOrder}{whether the order of the mediators is randomised before
the start of each iteration}

\item{subSampling}{whether to consider only a part of the variables at each
iteration, similar to how random forest decorrelates its trees. If true,
increase the stableLag parameter.}

\item{cutoff}{a cutoff value for selection: variables are selected if they
display a selection rate higher than this value. Only relevant when multiple
starts are specified. Can also be specified post-hoc using
\code{\link{setCutoff}}.}

\item{parallel}{whether to run the multiple starts in parallel}

\item{nCores}{how many threads (cores) to use for parallel processing}

\item{verbose}{whether to display iteration-based output and information.
Does not work with parallel processing.}

\item{progressBar}{whether to display a progress bar. Only relevant when
multiple starts are specified.}

\item{...}{parameters passed to decisionFunction}
}
\value{
an object of class \code{cmf}. See \code{\link{cmf-methods}}
}
\description{
This function performs CMF on a set of potential mediators, given an input
and an output variable. It selects those mediators that are deemed relevant
by a default or a user-defined decision function, *conditional* on the other
mediators in the model. By doing this cyclically, the algorithm converges to
the best fitting subset of mediators.
}
\details{
Available decision functions (arguments = defaultvalue):

 - \code{corMinusPartCor} (\code{p.value} = 0.05): Test for xy correlation minus the
 partial xy.m correlation

 - \code{prodCoef} (\code{p.value} = 0.05): Test for the product of coefficients, Sobel
 test
}
