% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmfast.R
\name{cmfast}
\alias{cmfast}
\title{Fast Coordinate-wise Mediation Filter backend}
\usage{
cmfast(x, M, y, decisionFunction = "prodcoef", nStarts, nCores, cutoff,
  maxIter, stableLag, p.value, pb)
}
\description{
Fast Coordinate-wise Mediation Filter backend
}
\keyword{internal}
