% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmfslow.R
\name{cmfslow}
\alias{cmfslow}
\title{R-based Coordinate-wise Mediation Filter backend}
\usage{
cmfslow(x, M, y, decisionFunction, nStarts, nCores, cutoff, maxIter, stableLag,
  ...)
}
\description{
R-based Coordinate-wise Mediation Filter backend
}
\keyword{internal}
